requirements.txt
Django==4.2.7
djangorestframework==3.14.0
sympy==1.12.1
numpy>=1.26.0
matplotlib>=3.8.0
Pillow>=10.1.0
setuptools>=68.0.0
wheel>=0.41.0
xhtml2pdf>=0.2.15
reportlab>=4.0
requests>=2.31.0
cerebras_cloud_sdk>=0.1.0
pytest>=8.0.0
pytest-django>=4.8.0
pytest-cov>=5.0.0